sge_api/requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
attrs==23.1.0
click==8.1.3
colorama==0.4.6
Django==4.2.1
Flask==2.2.0
Flask-MySQLdb==1.0.1
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mysql-connector-python==8.0.25
mysqlclient==2.1.1
protobuf==4.23.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.16
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
Werkzeug==2.3.4
